requirements.txt
# Core Dependencies
ccxt>=4.1.0
pandas>=2.0.0
numpy>=1.24.0
asyncio>=3.4.3
websockets>=11.0.3
aiohttp>=3.8.0

# Machine Learning
scikit-learn>=1.3.0
tensorflow>=2.13.0
torch>=2.0.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Data Processing
ta>=0.10.2
yfinance>=0.2.12
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.28.0
redis>=4.5.0

# API and Web
fastapi>=0.103.0
uvicorn>=0.23.0
requests>=2.31.0

# Notifications
python-telegram-bot>=20.5
smtplib
email-validator>=2.0.0

# Monitoring and Logging
prometheus-client>=0.17.0
structlog>=23.1.0
rich>=13.5.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
schedule>=1.2.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0